+++ src/test/kotlin/com/cmgapps/license/model/LicenseIdShould.kt
@@ -17,10 +17,13 @@ import org.hamcrest.Matchers.oneOf
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
+import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.EnumSource
+import org.junit.jupiter.params.provider.MethodSource
 import org.junit.jupiter.params.provider.ValueSource
+import java.util.stream.Stream
 
-class LicenseShould {
+class LicenseIdShould {
     @ParameterizedTest
     @EnumSource(
         LicenseId::class,
@@ -123,5 +126,28 @@ class LicenseShould {
     @Test
     fun `get empty text for UNKNOW license id`() {
         assertThat(LicenseId.UNKNOWN.text, `is`(emptyString()))
+    }
+
+    @ParameterizedTest(name = "${ParameterizedTest.DISPLAY_NAME_PLACEHOLDER} - LicenseId = {0}")
+    @MethodSource("provideLicenseIdMapping")
+    fun `map licenses to name and url`(id: LicenseId, size: Int) {
+        assertThat(LicenseId.map.filterValues { it == id }.size, `is`(size))
+    }
+
+    companion object {
+        @JvmStatic
+        private fun provideLicenseIdMapping(): Stream<Arguments> = Stream.of(
+            Arguments.of(LicenseId.APACHE, 12),
+            Arguments.of(LicenseId.BSD_2, 5),
+            Arguments.of(LicenseId.BSD_3, 4),
+            Arguments.of(LicenseId.EPL_2, 6),
+            Arguments.of(LicenseId.GPL_2, 7),
+            Arguments.of(LicenseId.GPL_3, 7),
+            Arguments.of(LicenseId.LGPL_2_1, 7),
+            Arguments.of(LicenseId.LGPL_3, 7),
+            Arguments.of(LicenseId.MIT, 4),
+            Arguments.of(LicenseId.MPL_2, 6),
+            Arguments.of(LicenseId.CDDL, 4),
+        )
     }
 }